+++ query-connector/src/app/designSystem/checkbox/checkbox.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/_variables" as *;
+@use "../../../styles/variables" as *;
 
 .checkbox {
   &.label,
