+++ query-connector/src/app/designSystem/drawer/drawer.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/_variables" as *;
+@use "../../../styles/_variables" as *;
 
 .drawer {
   position: fixed;
