+++ query-connector/src/app/designSystem/footer/footer.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/_variables" as *;
+@use "../../../styles/_variables" as *;
 
 .footerContainer {
   flex: 0 1 $footer-content-height;
