+++ query-connector/src/app/designSystem/footer/footer.tsx
