+++ query-connector/src/app/designSystem/header/header.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/_variables" as *;
+@use "../../../styles/_variables" as *;
 
 .dataUsagePolicyButton {
   cursor: pointer;
