+++ query-connector/src/app/designSystem/searchField/searchField.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/_variables" as *;
+@use "../../../styles/_variables" as *;
 
 .searchField {
   border-radius: 0.25rem;
@@ -6,8 +6,7 @@
   height: 3%;
   padding: 0.75rem;
   padding-left: 2.4rem;
-  background: url("../../../../styles/assets/search.svg") center / contain
-    no-repeat;
+  background: url("../../../styles/assets/search.svg") center / contain no-repeat;
   background-position: 12px 12px;
   float: none;
 }
