+++ query-connector/src/app/designSystem/sideNav/sidenav.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/_variables" as *;
+@use "../../../styles/_variables" as *;
 
 .container {
   background-color: $white !important;
