+++ query-connector/src/app/designSystem/table/table.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../styles/_variables" as *;
+@use "../../../styles/_variables" as *;
 
 .table {
   display: flex;
