+++ query-connector/src/app/query/components/CustomizeQuery.tsx
@@ -4,14 +4,14 @@ import React, { useState, useEffect } from "react";
 import { Button } from "@trussworks/react-uswds";
 import { DibbsConceptType, DibbsValueSet } from "../../constants";
 import { QueryResponse } from "@/app/query-service";
-import LoadingView from "./LoadingView";
-import { showToastConfirmation } from "../designSystem/toast/Toast";
+import LoadingView from "../../designSystem/LoadingView";
+import { showToastConfirmation } from "../../designSystem/toast/Toast";
 import styles from "./customizeQuery/customizeQuery.module.scss";
 import CustomizeQueryAccordionHeader from "./customizeQuery/CustomizeQueryAccordionHeader";
 import CustomizeQueryAccordionBody from "./customizeQuery/CustomizeQueryAccordionBody";
-import Accordion from "../designSystem/Accordion";
+import Accordion from "../../designSystem/Accordion";
 import CustomizeQueryNav from "./customizeQuery/CustomizeQueryNav";
-import Backlink from "./backLink/Backlink";
+import Backlink from "../../designSystem/backLink/Backlink";
 import { RETURN_LABEL } from "./stepIndicator/StepIndicator";
 import { generateValueSetGroupingsByDibbsConceptType } from "@/app/utils/valueSetTranslation";
 import { CustomUserQuery } from "@/app/query-building";
