+++ query-connector/src/app/query/components/resultsView/ResultsViewTable.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Accordion from "../../designSystem/Accordion";
+import Accordion from "../../../designSystem/Accordion";
 import styles from "./resultsView.module.scss";
 import ResultsViewAccordionBody from "./ResultsViewAccordionBody";
 import { ResultsViewAccordionItem } from "../ResultsView";
