+++ query-connector/src/app/query/components/resultsView/resultsView.module.scss
@@ -32,8 +32,7 @@
 }
 
 .accordionHeading {
-  font-family: "Source Sans Pro Web", "Helvetica Neue", "Helvetica", "Roboto",
-    "Arial", "sans-serif" !important;
+  font-family: "Source Sans Pro Web", "Helvetica Neue", "Helvetica", "Roboto", "Arial", "sans-serif" !important;
   font-size: 1.25rem !important;
   line-height: 1.1 !important;
 }
