+++ query-connector/src/app/query/components/resultsView/tableComponents/ImmunizationTable.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Table from "@/app/query/designSystem/table/Table";
+import Table from "@/app/designSystem/table/Table";
 import { Immunization } from "fhir/r4";
 import {
   formatDate,
