+++ query-connector/src/app/query/components/resultsView/tableComponents/ObservationTable.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Table from "@/app/query/designSystem/table/Table";
+import Table from "@/app/designSystem/table/Table";
 import { Observation } from "fhir/r4";
 import { formatCodeableConcept, formatDate } from "../../../../format-service";
 import { checkIfSomeElementWithPropertyExists } from "./utils";
